handle import of multiple files with the same name (fix #5)

+++ src/components/ProjectContext.tsx
@@ -77,10 +77,34 @@ const appStateReducer = (state, action) => {
       let newFiles = state.projectData.entries[entryIndex].files;
       for (const file of fileList) {
         try {
-          const destination = path.join(state.folderPath, file.name);
-          copyFileSync(file.path, destination);
-          console.log(`${file.path} was copied to ${destination}`);
-          newFiles = [...newFiles, { title: file.name }];
+          let saveFile = true;
+          let destination = path.join(state.folderPath, file.name);
+          let newName = file.name;
+
+          if (fs.existsSync(destination)) {
+            saveFile = window.confirm(
+              `A file with name ${newName} has already been imported. Do you want to import this file anyway, with a modified name?`
+            );
+
+            let i = 1;
+            do {
+              const parts = file.name.split('.');
+              const base = parts.slice(0, -1).join('');
+              const extension = parts.slice(-1)[0];
+              newName = `${base} (${i}).${extension}`;
+
+              destination = path.join(state.folderPath, newName);
+              console.log('Trying new name:', newName);
+
+              i += 1;
+            } while (fs.existsSync(destination));
+          }
+
+          if (saveFile) {
+            copyFileSync(file.path, destination);
+            console.log(`${file.path} was copied to ${destination}`);
+            newFiles = [...newFiles, { title: newName }];
+          }
         } catch (e) {
           console.log('Error', e.stack);
           console.log('Error', e.name);
